requirements.txt
datasets
transformers
trl>=0.8.0
peft
torch --index-url https://download.pytorch.org/whl/cu118
accelerate
bitsandbytes
google-cloud-aiplatform
google-cloud-logging
gcsfs
tensorboard
markupsafe==2.0.1
torchvision --index-url https://download.pytorch.org/whl/cu118
